Workflow server CD trigger
