+++ ee/codegen/package.json
@@ -40,6 +40,7 @@
   },
   "devDependencies": {
     "@octokit/auth-app": "^7.1.3",
+    "@octokit/rest": "^19.0.13",
     "@types/lodash": "^4.17.13",
     "@types/node": "^22.10.2",
     "@typescript-eslint/eslint-plugin": "^8.9.0",
